+++ CHANGELOG.md
@@ -15,8 +15,11 @@ You can also check the [release page](https://github.com/visualize-admin/visuali
 - Added migrations of chart configs (breaking changes to schema will not break existing charts)
 - Tech
   - Diminish size of Javascript that is shipped to the client. Should improve page load performance
-- Better formatting of numbers inside chart tooltips (for example for currencies)
-
+- Charts
+  - Better formatting of numbers inside chart tooltips (for example for currencies)
+- Filters
+  - Initially, left filters are ordered so that dimension with hierarchy are at the top. Additionally, the topmost hierarchical value with a value is selected.
+- 
 ## [3.7.11] - 2022-09-09
 
 - 🚀 Maps